requirements.txt
Flask==1.1.2
gunicorn==20.0.4
line-bot-sdk==1.16.
mccabe==0.6.1
mecab-python3==0.996.5
requests==2.23.0

